HoshinoBot/requirements.txt
nonebot[scheduler]~=1.6.0
aiocqhttp>=1.2
lxml>=4.4.1
pytz>=2019.3
requests>=2.22.0
sogou_tr_free>=0.0.6
zhconv>=1.4.0
Pillow>=6.2.1
TwitterAPI>=2.5.10
matplotlib>=3.2.0
numpy>=1.18.0
beautifulsoup4>=4.9.0
pygtrie>=2.0
tinydb>=4.0
aiohttp>=3.6
sqlitedict~=1.6.0
pyyaml
pixivpy
peewee
pandas
kth_timeoutdecorator
nest_asyncio
thulac
httpx
ujson
msgpack
logzero
python-Levenshtein
